SIG-3592 Split signal context serializers/views from signal ones
