+++ api/app/signals/apps/api/fields/__init__.py
@@ -2,6 +2,7 @@
 # Copyright (C) 2019 - 2021 Gemeente Amsterdam
 from signals.apps.api.fields.attachment import (
     PrivateSignalAttachmentLinksField,
+    PrivateSignalAttachmentRelatedField,
     PublicSignalAttachmentLinksField
 )
 from signals.apps.api.fields.category import (
@@ -9,10 +10,12 @@ from signals.apps.api.fields.category import (
     CategoryHyperlinkedRelatedField,
     PrivateCategoryHyperlinkedIdentityField
 )
+from signals.apps.api.fields.extra_properties import SignalExtraPropertiesField
 from signals.apps.api.fields.note import NoteHyperlinkedIdentityField
 from signals.apps.api.fields.signal import (
     PrivateSignalLinksField,
     PrivateSignalLinksFieldWithArchives,
+    PrivateSignalWithContextLinksField,
     PublicSignalLinksField
 )
 from signals.apps.api.fields.stored_signal_filter import StoredSignalFilterLinksField
@@ -21,11 +24,14 @@ __all__ = [
     'CategoryHyperlinkedIdentityField',
     'CategoryHyperlinkedRelatedField',
     'NoteHyperlinkedIdentityField',
-    'PrivateSignalLinksFieldWithArchives',
+    'PrivateCategoryHyperlinkedIdentityField',
+    'PrivateSignalAttachmentLinksField',
+    'PrivateSignalAttachmentRelatedField',
     'PrivateSignalLinksField',
+    'PrivateSignalLinksFieldWithArchives',
+    'PrivateSignalWithContextLinksField',
     'PublicSignalLinksField',
     'PublicSignalAttachmentLinksField',
-    'PrivateSignalAttachmentLinksField',
+    'SignalExtraPropertiesField',
     'StoredSignalFilterLinksField',
-    'PrivateCategoryHyperlinkedIdentityField',
 ]
