+++ api/app/signals/apps/api/views/signal.py
@@ -33,10 +33,6 @@ from signals.apps.api.serializers import (
     SignalGeoSerializer,
     SignalIdListSerializer
 )
-from signals.apps.api.serializers.signal import (
-    SignalContextReporterSerializer,
-    SignalContextSerializer
-)
 from signals.apps.api.views._base import PublicSignalGenericViewSet
 from signals.apps.signals import workflow
 from signals.apps.signals.models import Signal
@@ -295,51 +291,3 @@ class SignalPromotedToParentViewSet(GenericViewSet, mixins.ListModelMixin):
     filterset_class = SignalPromotedToParentFilter
 
     queryset = Signal.objects.prefetch_related('children').only('id').filter(children__isnull=False).order_by('-id')
-
-
-class SignalContextViewSet(mixins.RetrieveModelMixin, GenericViewSet):
-    serializer_class = SignalContextSerializer
-    serializer_detail_class = SignalContextSerializer
-
-    pagination_class = HALPagination
-
-    authentication_classes = (JWTAuthBackend,)
-    permission_classes = (SIAPermissions,)
-
-    def get_queryset(self, *args, **kwargs):
-        return Signal.objects.filter_for_user(user=self.request.user)
-
-    def geography(self, request, pk=None):
-        paginator = LinkHeaderPagination(page_query_param='geopage', page_size=4000)
-        page = paginator.paginate_queryset(Signal.objects.none(), self.request, view=self)
-        if page is not None:
-            serializer = SignalGeoSerializer(page, many=True, context=self.get_serializer_context())
-            return paginator.get_paginated_response(serializer.data)
-
-        serializer = SignalGeoSerializer(page, many=True, context=self.get_serializer_context())
-        return Response(serializer.data)
-
-    def reporter(self, request, pk=None):
-        signal = self.get_object()
-        if signal.reporter.email:
-            signals_for_reporter_qs = Signal.objects.select_related(
-                'category_assignment',
-                'category_assignment__category',
-                'status',
-            ).prefetch_related(
-                'feedback',
-                'category_assignment__category__departments'
-            ).filter_reporter(
-                email=signal.reporter.email
-            )
-        else:
-            signals_for_reporter_qs = Signal.objects.none()
-
-        page = self.paginate_queryset(signals_for_reporter_qs)
-        if page is not None:
-            serializer = SignalContextReporterSerializer(page, many=True, context=self.get_serializer_context())
-            return self.get_paginated_response(serializer.data)
-
-        serializer = SignalContextReporterSerializer(signals_for_reporter_qs, many=True,
-                                                     context=self.get_serializer_context())
-        return Response(serializer.data)
